server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cachetools==6.2.4
cbor2==5.8.0
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
constantly==23.10.4
cryptography==46.0.3
cssselect2==0.8.0
daphne==4.2.1
Django==6.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-redis==6.0.0
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
freetype-py==2.5.1
frozenlist==1.8.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
html5lib==1.1
httplib2==0.31.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
kombu==5.6.1
lxml==6.0.2
msgpack==1.1.2
multidict==6.7.0
oauthlib==3.3.1
oscrypto==1.3.0
packaging==25.0
pillow==12.1.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==3.0
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.3.1
pypdf==6.6.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
reportlab==4.4.9
requests==2.32.5
requests-oauthlib==2.0.0
rlPyCairo==0.4.0
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
svglib==1.6.0
tinycss2==1.5.1
twilio==9.9.1
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritools==6.0.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
wheel==0.45.1
xhtml2pdf==0.2.17
yarl==1.22.0
zope.interface==8.2
